Standardize usage tracking and response formats across all providers
- Add common TokenUsage type with optional fields and provider-specific tokens
- Add unified Timestamp type with conversion utilities for Unix/ISO formats
- Add ResponseFormat enum for OpenAI-compatible structured outputs
- Update all providers to use common usage tracking instead of provider-specific types
- Consolidate duplicate response structures in openai_format module
- Fix usage token access methods across Groq, Mistral, OpenRouter conversions
- Add comprehensive usage merging and arithmetic operations
- Improve type consistency for timestamps and token counting

+++ crates/mistral-ox/src/usage.rs
@@ -2,9 +2,9 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, Default, Serialize, Deserialize)]
 pub struct Usage {
-    pub prompt_tokens: u32,
-    pub completion_tokens: u32,
-    pub total_tokens: u32,
+    pub prompt_tokens: u64,
+    pub completion_tokens: u64,
+    pub total_tokens: u64,
 }
 
 impl Usage {
